feat(modularization): S-8 -- extract memory view + graph frontend to web/modules/memories.js
Moves the full memories UI section (~1243 lines in two blocks) from the
monolithic web/app.js into web/modules/memories.js as an ES module.

Covered sections:
- Memories (Tier System, search/filter, CRUD, daily log): lines 577-1660
- Memory Import overlay: lines 3861-4019

Exports: initMemories, loadMemAgents, loadMemStats, loadMemories,
         openEditMemory, loadDailyLog
DI pattern: openModal/closeModal injected via initMemories({...})
Local helpers: mainAgentId(), escapeHtml() (inlined)
Imports: showToast (toast.js), t (i18n.js)

Graph section (force-directed canvas, Obsidian-style): fully self-contained
canvas API, no external CDN lib -- no load-order concern.

app.js: added import + initMemories({ openModal, closeModal }) call;
removed both source blocks (app.js: 10606 -> 9352 lines).

Verified: node --check, npm run build, 2873 vitest tests passing,
Playwright headless smoke 0 JS errors (memories/tasks/agents/kanban pages).
