+++ web/app.js
@@ -11,6 +11,7 @@ import {
   avatarBust, agentApiName, populateAvatarGrid, loadAvailableModels,
 } from './modules/agents.js'
 import { initSchedules, loadSchedules, loadScheduleAgents, openEditSchedule, getScheduleCron } from './modules/schedules.js'
+import { initMemories, loadMemAgents, loadMemStats, loadMemories } from './modules/memories.js'
 
 // avatarBust() is imported from ./modules/agents.js (avatar epoch owned there).
 
@@ -386,6 +387,7 @@ initAgents({
 
 // Wire modal helpers into the schedules module.
 initSchedules({ openModal, closeModal })
+initMemories({ openModal, closeModal })
 
 // === Skills ===
 async function loadSkills(agentName) {
@@ -574,1100 +576,6 @@ document.getElementById('importSkillBtn').addEventListener('click', async () => 
   }
 })
 
-// ============================================================
-// ============================================================
-// === Memories (Tier System + Daily Log) ===
-// ============================================================
-
-const memList = document.getElementById('memList')
-const memEmpty = document.getElementById('memEmpty')
-const memStats = document.getElementById('memStats')
-const memSearchInput = document.getElementById('memSearchInput')
-const memModalOverlay = document.getElementById('memModalOverlay')
-
-let memSearchTimer = null
-let currentMemTier = 'hot'
-let currentLogDate = new Date().toISOString().split('T')[0]
-let logDates = []
-
-const tierLabels = { hot: '\u{1F525} Hot', warm: '\u{1F321}\uFE0F Warm', cold: '\u2744\uFE0F Cold', shared: '\u{1F517} Shared' }
-const tierColors = { hot: '#dc3c3c', warm: '#d97757', cold: '#6a9bcc', shared: '#9a8a30' }
-
-// Populate agent dropdowns from API
-async function loadMemAgents() {
-  try {
-    const res = await fetch('/api/schedules/agents')
-    const agents = await res.json()
-    const sel = document.getElementById('memAgentFilter')
-    const memSel = document.getElementById('memAgent')
-    sel.innerHTML = `<option value="">${t('memories.agent_all')}</option>`
-    memSel.innerHTML = ''
-    for (const a of agents) {
-      sel.innerHTML += `<option value="${a.name}">${a.label}</option>`
-      memSel.innerHTML += `<option value="${a.name}">${a.label}</option>`
-    }
-  } catch {}
-}
-
-// Agent filter change
-document.getElementById('memAgentFilter').addEventListener('change', () => {
-  if (currentMemTier === 'graph') {
-    loadMemoryGraph()
-  } else if (currentMemTier === 'log') {
-    loadDailyLog()
-  } else {
-    loadMemories()
-  }
-})
-
-// Search with debounce
-memSearchInput.addEventListener('input', () => {
-  clearTimeout(memSearchTimer)
-  memSearchTimer = setTimeout(loadMemories, 300)
-})
-
-// Enter to search immediately
-memSearchInput.addEventListener('keydown', (e) => {
-  if (e.key === 'Enter') {
-    clearTimeout(memSearchTimer)
-    loadMemories()
-  }
-})
-
-// Tab switching
-document.getElementById('memTabs').addEventListener('click', (e) => {
-  const tab = e.target.closest('.mem-tab')
-  if (!tab) return
-  document.querySelectorAll('.mem-tab').forEach(t => t.classList.remove('active'))
-  tab.classList.add('active')
-  currentMemTier = tab.dataset.tier
-
-  const isLog = currentMemTier === 'log'
-  const isGraph = currentMemTier === 'graph'
-  document.getElementById('memTierView').hidden = isLog || isGraph
-  document.getElementById('memLogView').hidden = !isLog
-  document.getElementById('memGraphView').hidden = !isGraph
-
-  if (isGraph) {
-    loadMemoryGraph()
-  } else if (isLog) {
-    loadDailyLog()
-  } else {
-    loadMemories()
-  }
-})
-
-// Add memory button
-document.getElementById('memAddBtn').addEventListener('click', () => {
-  document.getElementById('memModalTitle').textContent = t('memories.modal.title_new')
-  document.getElementById('memContent').value = ''
-  document.getElementById('memTier').value = (currentMemTier === 'log' || currentMemTier === 'graph') ? 'warm' : currentMemTier
-  document.getElementById('memKeywords').value = ''
-  document.getElementById('memEditId').value = ''
-  openModal(memModalOverlay)
-  setTimeout(() => document.getElementById('memContent').focus(), 200)
-})
-
-// Close memory modal
-document.getElementById('memModalClose').addEventListener('click', () => closeModal(memModalOverlay))
-memModalOverlay.addEventListener('click', (e) => { if (e.target === memModalOverlay) closeModal(memModalOverlay) })
-
-// Save memory (create or edit)
-document.getElementById('saveMemBtn').addEventListener('click', async () => {
-  const content = document.getElementById('memContent').value.trim()
-  if (!content) { document.getElementById('memContent').focus(); return }
-
-  const editId = document.getElementById('memEditId').value
-  const tier = document.getElementById('memTier').value
-  const agentId = document.getElementById('memAgent').value
-  const keywords = document.getElementById('memKeywords').value.trim()
-
-  try {
-    if (editId) {
-      await fetch(`/api/memories/${editId}`, {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ content, tier, agent_id: agentId, keywords }),
-      })
-      showToast(t('memories.toast.updated'))
-    } else {
-      await fetch('/api/memories', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ agent_id: agentId, content, tier, keywords }),
-      })
-      showToast(t('memories.toast.created'))
-    }
-    closeModal(memModalOverlay)
-    loadMemories()
-    loadMemStats()
-  } catch {
-    showToast(t('common.error_save'))
-  }
-})
-
-async function loadMemStats() {
-  try {
-    const res = await fetch('/api/memories/stats')
-    const stats = await res.json()
-    const embCount = stats.withEmbedding || 0
-    const embPct = stats.total > 0 ? Math.round(embCount / stats.total * 100) : 0
-    memStats.innerHTML = `
-      <div class="stat-card"><div class="stat-value">${stats.total}</div><div class="stat-label">${t('memories.stat.total')}</div></div>
-      ${Object.entries(stats.byTier || {}).map(([tier, count]) =>
-        `<div class="stat-card"><div class="stat-value" style="color:${tierColors[tier] || 'var(--accent)'}">${count}</div><div class="stat-label">${tierLabels[tier] || tier}</div></div>`
-      ).join('')}
-      <div class="stat-card"><div class="stat-value">${embCount}</div><div class="stat-label">${t('memories.stat.vectors_pct', { pct: embPct })}</div></div>
-      <button class="btn-secondary btn-compact" id="memBackfillBtn" style="margin-left:auto;font-size:11px;padding:6px 12px;align-self:center">${t('memories.stat.vectors_btn')}</button>
-    `
-    document.getElementById('memBackfillBtn')?.addEventListener('click', async () => {
-      const btn = document.getElementById('memBackfillBtn')
-      if (btn) { btn.textContent = t('memories.stat.vectors_gen'); btn.disabled = true }
-      try {
-        const r = await fetch('/api/memories/backfill', { method: 'POST' })
-        const data = await r.json()
-        showToast(t('memories.toast.vector_count', { count: data.count }))
-        loadMemStats()
-      } catch { showToast(t('memories.toast.vector_error')) }
-    })
-  } catch (err) {
-    console.error('Stats hiba:', err)
-  }
-}
-
-async function loadMemories() {
-  if (currentMemTier === 'log' || currentMemTier === 'graph') return
-  const q = memSearchInput.value.trim()
-  const agent = document.getElementById('memAgentFilter').value
-  const searchMode = document.getElementById('memSearchMode')?.value || 'hybrid'
-  const params = new URLSearchParams()
-  if (q) {
-    params.set('q', q)
-    params.set('mode', searchMode)
-  }
-  if (agent) params.set('agent', agent)
-  if (currentMemTier) params.set('tier', currentMemTier)
-  params.set('limit', '50')
-
-  try {
-    const res = await fetch(`/api/memories?${params}`)
-    const memories = await res.json()
-    renderMemories(memories)
-  } catch (err) {
-    console.error('Memória betöltés hiba:', err)
-  }
-}
-
-function renderMemories(memories) {
-  memList.innerHTML = ''
-  memEmpty.hidden = memories.length > 0
-
-  for (const mem of memories) {
-    const item = document.createElement('div')
-    item.className = 'mem-item'
-
-    const tier = mem.tier || mem.category || 'warm'
-    const tierBadge = tierLabels[tier] || tier
-    const badgeClass = 'badge-' + tier
-    const shortContent = mem.content.length > 120 ? mem.content.slice(0, 120) + '...' : mem.content
-    const agentLabel = mem.agent_id || mainAgentId()
-
-    // Build keywords HTML
-    let keywordsHtml = ''
-    if (mem.keywords) {
-      const kws = typeof mem.keywords === 'string' ? mem.keywords.split(',').map(k => k.trim()).filter(Boolean) : mem.keywords
-      if (kws.length > 0) {
-        keywordsHtml = `<div class="mem-keywords">${kws.map(k => `<span class="mem-keyword-tag">${escapeHtml(k)}</span>`).join('')}</div>`
-      }
-    }
-
-    item.innerHTML = `
-      <div class="mem-item-header">
-        <span class="badge ${badgeClass}">${tierBadge}</span>
-        <span class="mem-agent-badge">${escapeHtml(agentLabel)}</span>
-        <span class="mem-date">${escapeHtml(mem.created_label || '')}</span>
-        ${typeof mem.salience === 'number' ? `<span class="mem-salience" title="Relevancia ertek">S: ${mem.salience.toFixed(2)}</span>` : ''}
-      </div>
-      <div class="mem-content-short">${escapeHtml(shortContent)}</div>
-      <div class="mem-content-full">${escapeHtml(mem.content)}</div>
-      ${keywordsHtml}
-      <div class="mem-item-footer">
-        <button class="btn-secondary" data-edit-memid="${mem.id}" style="padding:6px 14px; font-size:12px;">${t('common.btn.edit')}</button>
-        <button class="btn-danger" data-memid="${mem.id}" style="padding:6px 14px; font-size:12px;">${t('common.btn.delete')}</button>
-      </div>
-    `
-
-    // Toggle expand
-    item.addEventListener('click', (e) => {
-      if (e.target.closest('.btn-danger') || e.target.closest('.btn-secondary')) return
-      item.classList.toggle('expanded')
-    })
-
-    // Edit
-    const editBtn = item.querySelector('[data-edit-memid]')
-    editBtn.addEventListener('click', (e) => {
-      e.stopPropagation()
-      document.getElementById('memModalTitle').textContent = t('memories.modal.title_edit')
-      document.getElementById('memContent').value = mem.content
-      document.getElementById('memTier').value = tier
-      document.getElementById('memKeywords').value = mem.keywords || ''
-      document.getElementById('memEditId').value = mem.id
-      if (mem.agent_id) document.getElementById('memAgent').value = mem.agent_id
-      openModal(memModalOverlay)
-    })
-
-    // Delete
-    const delBtn = item.querySelector('.btn-danger')
-    delBtn.addEventListener('click', async (e) => {
-      e.stopPropagation()
-      if (!confirm('Biztosan torlod ezt az emleket?')) return
-      try {
-        await fetch(`/api/memories/${mem.id}`, { method: 'DELETE' })
-        showToast(t('memories.toast.deleted'))
-        loadMemories()
-        loadMemStats()
-      } catch {
-        showToast(t('common.error_delete'))
-      }
-    })
-
-    memList.appendChild(item)
-  }
-}
-
-// === Memory Graph (Force-directed, Obsidian-style) ===
-
-let graphNodes = []
-let graphEdges = []
-let graphSim = null
-let graphCanvas = null
-let graphCtx = null
-let graphDragging = null
-let graphHover = null
-let graphSelectedNode = null
-let graphSearchQuery = ''
-
-// Zoom & pan state
-let graphZoom = 1
-let graphPanX = 0
-let graphPanY = 0
-let graphPanning = false
-let graphPanStartX = 0
-let graphPanStartY = 0
-let graphZoomIndicatorTimer = null
-
-// Edge animation
-let graphAnimFrame = 0
-
-const GRAPH_TIER_COLORS = {
-  hot: '#dc3c3c',
-  warm: '#d97757',
-  cold: '#6a9bcc',
-  shared: '#b0a040',
-}
-
-const GRAPH_TIER_BG = {
-  hot: 'rgba(220, 60, 60, 0.06)',
-  warm: 'rgba(217, 119, 87, 0.06)',
-  cold: 'rgba(106, 155, 204, 0.06)',
-  shared: 'rgba(176, 160, 64, 0.06)',
-}
-
-function screenToWorld(sx, sy) {
-  return { x: (sx - graphPanX) / graphZoom, y: (sy - graphPanY) / graphZoom }
-}
-
-function worldToScreen(wx, wy) {
-  return { x: wx * graphZoom + graphPanX, y: wy * graphZoom + graphPanY }
-}
-
-async function loadMemoryGraph() {
-  const agent = document.getElementById('memAgentFilter').value
-  const params = new URLSearchParams()
-  if (agent) params.set('agent', agent)
-  params.set('limit', '200')
-
-  try {
-    const res = await fetch(`/api/memories?${params}`)
-    const memories = await res.json()
-
-    const emptyEl = document.getElementById('graphEmpty')
-    if (!memories || memories.length === 0) {
-      emptyEl.hidden = false
-      document.getElementById('memGraphCanvas').hidden = true
-      return
-    }
-    emptyEl.hidden = true
-    document.getElementById('memGraphCanvas').hidden = false
-
-    // Reset zoom/pan on new data load
-    graphZoom = 1
-    graphPanX = 0
-    graphPanY = 0
-    graphSelectedNode = null
-    hideGraphPanel()
-
-    buildGraph(memories)
-    startGraphSimulation()
-  } catch (err) {
-    console.error('Gráf betöltés hiba:', err)
-  }
-}
-
-function buildGraph(memories) {
-  graphNodes = []
-  graphEdges = []
-
-  const canvas = document.getElementById('memGraphCanvas')
-  const rect = canvas.parentElement.getBoundingClientRect()
-  const dpr = window.devicePixelRatio || 1
-  canvas.width = rect.width * dpr
-  canvas.height = rect.height * dpr
-  canvas.style.width = rect.width + 'px'
-  canvas.style.height = rect.height + 'px'
-  graphCanvas = canvas
-  graphCtx = canvas.getContext('2d')
-  graphCtx.setTransform(dpr, 0, 0, dpr, 0, 0)
-
-  const w = rect.width
-  const h = rect.height
-
-  // Create nodes from memories
-  for (const mem of memories) {
-    const keywords = (mem.keywords || '').split(',').map(k => k.trim()).filter(Boolean)
-    const label = mem.content.slice(0, 25).replace(/\n/g, ' ') + (mem.content.length > 25 ? '...' : '')
-    graphNodes.push({
-      id: mem.id,
-      x: w / 2 + (Math.random() - 0.5) * w * 0.6,
-      y: h / 2 + (Math.random() - 0.5) * h * 0.6,
-      vx: 0,
-      vy: 0,
-      radius: 6,
-      connectionCount: 0,
-      label: label,
-      tier: mem.tier || mem.category || 'warm',
-      agent: mem.agent_id || mainAgentId(),
-      keywords: keywords,
-      mem: mem,
-      searchMatch: true,
-    })
-  }
-
-  // Create edges based on shared keywords
-  for (let i = 0; i < graphNodes.length; i++) {
-    for (let j = i + 1; j < graphNodes.length; j++) {
-      const a = graphNodes[i]
-      const b = graphNodes[j]
-      const shared = a.keywords.filter(k => b.keywords.includes(k))
-      if (shared.length > 0) {
-        graphEdges.push({ source: i, target: j, strength: shared.length })
-        a.connectionCount += shared.length
-        b.connectionCount += shared.length
-      }
-      // Also connect same-agent same-tier with low probability
-      if (a.agent === b.agent && a.tier === b.tier && Math.random() < 0.3) {
-        graphEdges.push({ source: i, target: j, strength: 0.5 })
-        a.connectionCount += 0.5
-        b.connectionCount += 0.5
-      }
-    }
-  }
-
-  // Set node radius based on connection count
-  for (const node of graphNodes) {
-    node.radius = 5 + Math.min(Math.sqrt(node.connectionCount) * 2.5, 14)
-  }
-
-  // Ensure controls hint and zoom indicator exist
-  const graphView = document.getElementById('memGraphView')
-  if (!graphView.querySelector('.graph-controls-hint')) {
-    const hint = document.createElement('div')
-    hint.className = 'graph-controls-hint'
-    hint.innerHTML = 'Scroll: zoom | Drag: move nodes<br>Click: details | Dbl-click: edit'
-    graphView.appendChild(hint)
-  }
-  if (!graphView.querySelector('.graph-zoom-indicator')) {
-    const zi = document.createElement('div')
-    zi.className = 'graph-zoom-indicator'
-    zi.id = 'graphZoomIndicator'
-    graphView.appendChild(zi)
-  }
-}
-
-function simulateGraphStep(damping) {
-  const w = graphCanvas.width / (window.devicePixelRatio || 1)
-  const h = graphCanvas.height / (window.devicePixelRatio || 1)
-  const nodes = graphNodes
-
-  const tierCenters = {}
-  for (const node of nodes) {
-    if (!tierCenters[node.tier]) tierCenters[node.tier] = { x: 0, y: 0, count: 0 }
-    tierCenters[node.tier].x += node.x
-    tierCenters[node.tier].y += node.y
-    tierCenters[node.tier].count++
-  }
-  for (const tier of Object.keys(tierCenters)) {
-    tierCenters[tier].x /= tierCenters[tier].count
-    tierCenters[tier].y /= tierCenters[tier].count
-  }
-  for (const node of nodes) {
-    const tc = tierCenters[node.tier]
-    if (tc) {
-      node.vx += (tc.x - node.x) * 0.0005
-      node.vy += (tc.y - node.y) * 0.0005
-    }
-  }
-
-  for (let i = 0; i < nodes.length; i++) {
-    for (let j = i + 1; j < nodes.length; j++) {
-      let dx = nodes[j].x - nodes[i].x
-      let dy = nodes[j].y - nodes[i].y
-      let dist = Math.sqrt(dx * dx + dy * dy) || 1
-      let force = 800 / (dist * dist)
-      let fx = (dx / dist) * force
-      let fy = (dy / dist) * force
-      nodes[i].vx -= fx
-      nodes[i].vy -= fy
-      nodes[j].vx += fx
-      nodes[j].vy += fy
-    }
-  }
-
-  for (const edge of graphEdges) {
-    const a = nodes[edge.source]
-    const b = nodes[edge.target]
-    let dx = b.x - a.x
-    let dy = b.y - a.y
-    let dist = Math.sqrt(dx * dx + dy * dy) || 1
-    let force = (dist - 80) * 0.005 * edge.strength
-    let fx = (dx / dist) * force
-    let fy = (dy / dist) * force
-    a.vx += fx
-    a.vy += fy
-    b.vx -= fx
-    b.vy -= fy
-  }
-
-  for (const node of nodes) {
-    node.vx += (w / 2 - node.x) * 0.001
-    node.vy += (h / 2 - node.y) * 0.001
-  }
-
-  const maxV = 6
-  for (const node of nodes) {
-    if (node === graphDragging) continue
-    node.vx *= damping
-    node.vy *= damping
-    if (node.vx > maxV) node.vx = maxV; else if (node.vx < -maxV) node.vx = -maxV
-    if (node.vy > maxV) node.vy = maxV; else if (node.vy < -maxV) node.vy = -maxV
-    node.x += node.vx
-    node.y += node.vy
-    node.x = Math.max(-200, Math.min(w + 200, node.x))
-    node.y = Math.max(-200, Math.min(h + 200, node.y))
-  }
-}
-
-function startGraphSimulation() {
-  if (graphSim) cancelAnimationFrame(graphSim)
-
-  for (const node of graphNodes) {
-    node.vx = 0
-    node.vy = 0
-  }
-
-  const preSettleIterations = Math.min(250, 40 + graphNodes.length * 2)
-  for (let i = 0; i < preSettleIterations; i++) {
-    simulateGraphStep(0.88)
-  }
-
-  let frame = 0
-  const maxFrames = 60
-
-  function tick() {
-    if (frame > maxFrames) {
-      renderGraph()
-      return
-    }
-    frame++
-    graphAnimFrame = frame
-    simulateGraphStep(0.94 + (frame / maxFrames) * 0.05)
-    renderGraph()
-    graphSim = requestAnimationFrame(tick)
-  }
-
-  tick()
-}
-
-function renderGraph() {
-  const ctx = graphCtx
-  const dpr = window.devicePixelRatio || 1
-  const w = graphCanvas.width / dpr
-  const h = graphCanvas.height / dpr
-
-  ctx.clearRect(0, 0, w, h)
-
-  const cs = getComputedStyle(document.documentElement)
-  const borderColor = cs.getPropertyValue('--border').trim() || '#d1cfc5'
-  const textColor = cs.getPropertyValue('--text').trim() || '#141413'
-  const textMuted = cs.getPropertyValue('--text-muted').trim() || '#87867f'
-  const bgCard = cs.getPropertyValue('--bg-card').trim() || '#fff'
-  const bgColor = cs.getPropertyValue('--bg').trim() || '#faf9f5'
-  const isDark = document.documentElement.getAttribute('data-theme') === 'dark'
-
-  // === Dot grid background (drawn in screen space) ===
-  const gridSize = 20
-  const dotColor = borderColor
-  ctx.fillStyle = dotColor
-  ctx.globalAlpha = isDark ? 0.2 : 0.3
-  const offsetX = ((graphPanX % (gridSize * graphZoom)) + gridSize * graphZoom) % (gridSize * graphZoom)
-  const offsetY = ((graphPanY % (gridSize * graphZoom)) + gridSize * graphZoom) % (gridSize * graphZoom)
-  const scaledGrid = gridSize * graphZoom
-  if (scaledGrid > 4) {
-    for (let x = offsetX; x < w; x += scaledGrid) {
-      for (let y = offsetY; y < h; y += scaledGrid) {
-        ctx.beginPath()
-        ctx.arc(x, y, Math.max(0.5, graphZoom * 0.6), 0, Math.PI * 2)
-        ctx.fill()
-      }
-    }
-  }
-  ctx.globalAlpha = 1
-
-  // === Apply zoom/pan transform ===
-  ctx.save()
-  ctx.translate(graphPanX, graphPanY)
-  ctx.scale(graphZoom, graphZoom)
-
-  const hasSearch = graphSearchQuery.length > 0
-
-  // === Tier cluster backgrounds ===
-  const tierGroups = {}
-  for (const node of graphNodes) {
-    if (!tierGroups[node.tier]) tierGroups[node.tier] = []
-    tierGroups[node.tier].push(node)
-  }
-  for (const [tier, nodes] of Object.entries(tierGroups)) {
-    if (nodes.length < 2) continue
-    let cx = 0, cy = 0
-    for (const n of nodes) { cx += n.x; cy += n.y }
-    cx /= nodes.length
-    cy /= nodes.length
-    let maxDist = 0
-    for (const n of nodes) {
-      const d = Math.sqrt((n.x - cx) ** 2 + (n.y - cy) ** 2)
-      if (d > maxDist) maxDist = d
-    }
-    const radius = maxDist + 60
-    const grad = ctx.createRadialGradient(cx, cy, 0, cx, cy, radius)
-    const bgTier = GRAPH_TIER_BG[tier] || 'rgba(128,128,128,0.04)'
-    grad.addColorStop(0, bgTier)
-    grad.addColorStop(1, 'rgba(0,0,0,0)')
-    ctx.fillStyle = grad
-    ctx.globalAlpha = hasSearch ? 0.3 : 0.8
-    ctx.beginPath()
-    ctx.arc(cx, cy, radius, 0, Math.PI * 2)
-    ctx.fill()
-    ctx.globalAlpha = 1
-  }
-
-  // Build set of connected node indices for hovered/selected node
-  const connectedToActive = new Set()
-  const activeNode = graphHover || graphSelectedNode
-  if (activeNode) {
-    const activeIdx = graphNodes.indexOf(activeNode)
-    for (const edge of graphEdges) {
-      if (edge.source === activeIdx) connectedToActive.add(edge.target)
-      if (edge.target === activeIdx) connectedToActive.add(edge.source)
-    }
-  }
-
-  // === Draw edges (bezier curves with pulsing) ===
-  const time = Date.now() * 0.001
-  for (const edge of graphEdges) {
-    const a = graphNodes[edge.source]
-    const b = graphNodes[edge.target]
-
-    const isActiveEdge = activeNode && (a === activeNode || b === activeNode)
-    const searchFaded = hasSearch && (!a.searchMatch || !b.searchMatch)
-
-    // Edge thickness based on connection strength
-    const baseWidth = 0.5 + Math.min(edge.strength * 0.6, 2.5)
-
-    // Subtle pulse/breathe animation
-    const pulse = 0.85 + 0.15 * Math.sin(time * 1.5 + edge.source * 0.3 + edge.target * 0.7)
-
-    ctx.lineWidth = isActiveEdge ? baseWidth * 1.8 : baseWidth * pulse
-    ctx.strokeStyle = isActiveEdge ? GRAPH_TIER_COLORS[a === activeNode ? a.tier : b.tier] || borderColor : borderColor
-    ctx.globalAlpha = searchFaded ? 0.05 : (isActiveEdge ? 0.7 : (0.15 + Math.min(edge.strength * 0.1, 0.3)) * pulse)
-
-    // Bezier curve: midpoint offset perpendicular to the line
-    const mx = (a.x + b.x) / 2
-    const my = (a.y + b.y) / 2
-    const dx = b.x - a.x
-    const dy = b.y - a.y
-    const dist = Math.sqrt(dx * dx + dy * dy) || 1
-    const curvature = Math.min(dist * 0.15, 30)
-    // Perpendicular offset
-    const cpx = mx + (-dy / dist) * curvature
-    const cpy = my + (dx / dist) * curvature
-
-    ctx.beginPath()
-    ctx.moveTo(a.x, a.y)
-    ctx.quadraticCurveTo(cpx, cpy, b.x, b.y)
-    ctx.stroke()
-  }
-  ctx.globalAlpha = 1
-
-  // === Draw nodes ===
-  const fontSize = Math.max(8, Math.min(12, 10 / graphZoom))
-
-  for (let ni = 0; ni < graphNodes.length; ni++) {
-    const node = graphNodes[ni]
-    const color = GRAPH_TIER_COLORS[node.tier] || '#d97757'
-    const isHover = node === graphHover
-    const isSelected = node === graphSelectedNode
-    const isConnected = connectedToActive.has(ni)
-    const searchFaded = hasSearch && !node.searchMatch
-    const searchGlow = hasSearch && node.searchMatch
-
-    // Opacity
-    let nodeAlpha = 0.85
-    if (searchFaded) nodeAlpha = 0.12
-    else if (searchGlow) nodeAlpha = 1
-    else if (isHover || isSelected) nodeAlpha = 1
-    else if (activeNode && !isConnected) nodeAlpha = 0.35
-
-    // Glow effect for hover, selected, search match
-    if ((isHover || isSelected || searchGlow) && !searchFaded) {
-      ctx.shadowColor = color
-      ctx.shadowBlur = isHover ? 20 : (searchGlow ? 15 : 10)
-    }
-
-    // Connected nodes get subtle highlight
-    if (isConnected && !searchFaded) {
-      ctx.shadowColor = color
-      ctx.shadowBlur = 6
-    }
-
-    const r = isHover ? node.radius + 3 : (isSelected ? node.radius + 2 : node.radius)
-
-    // Node fill
-    ctx.fillStyle = color
-    ctx.globalAlpha = nodeAlpha
-    ctx.beginPath()
-    ctx.arc(node.x, node.y, r, 0, Math.PI * 2)
-    ctx.fill()
-
-    // Subtle border ring for selected
-    if (isSelected) {
-      ctx.strokeStyle = color
-      ctx.lineWidth = 2
-      ctx.globalAlpha = 0.6
-      ctx.beginPath()
-      ctx.arc(node.x, node.y, r + 4, 0, Math.PI * 2)
-      ctx.stroke()
-    }
-
-    ctx.shadowBlur = 0
-    ctx.shadowColor = 'transparent'
-
-    // === Always show label (pill/badge style) ===
-    if (!searchFaded || (searchFaded && nodeAlpha > 0.15)) {
-      const labelText = node.label
-      const labelFontSize = Math.max(7, Math.min(11, 9 / Math.max(graphZoom * 0.7, 0.5)))
-      ctx.font = (isHover || isSelected) ? `600 ${labelFontSize + 1}px -apple-system, sans-serif` : `500 ${labelFontSize}px -apple-system, sans-serif`
-      const textWidth = ctx.measureText(labelText).width
-      const pillW = textWidth + 10
-      const pillH = labelFontSize + 6
-      const pillX = node.x - pillW / 2
-      const pillY = node.y + r + 5
-
-      // Dark pill background
-      ctx.globalAlpha = searchFaded ? 0.08 : ((isHover || isSelected) ? 0.9 : 0.65)
-      ctx.fillStyle = isDark ? 'rgba(20,20,19,0.85)' : 'rgba(30,30,28,0.8)'
-      graphRoundRect(ctx, pillX, pillY, pillW, pillH, 3)
-      ctx.fill()
-
-      // White text
-      ctx.fillStyle = isDark ? '#e8e7e0' : '#faf9f5'
-      ctx.globalAlpha = searchFaded ? 0.1 : ((isHover || isSelected) ? 1 : 0.85)
-      ctx.textAlign = 'center'
-      ctx.textBaseline = 'middle'
-      ctx.fillText(labelText, node.x, pillY + pillH / 2)
-    }
-
-    ctx.globalAlpha = 1
-    ctx.textBaseline = 'alphabetic'
-  }
-
-  // Hover tooltip (richer than before)
-  if (graphHover && !graphSelectedNode) {
-    const node = graphHover
-    const tLabels = { hot: 'Hot', warm: 'Warm', cold: 'Cold', shared: 'Shared' }
-    const text = `${tLabels[node.tier] || node.tier} | ${node.agent}`
-    const kw = node.keywords.length > 0 ? node.keywords.join(', ') : ''
-    const conns = `${Math.round(node.connectionCount)} connections`
-
-    ctx.font = 'bold 11px -apple-system, sans-serif'
-    const tw = Math.max(ctx.measureText(text).width, kw ? ctx.measureText(kw).width : 0, ctx.measureText(conns).width) + 24
-    const th = kw ? 64 : 48
-    let tx = node.x - tw / 2
-    let ty = node.y - node.radius - th - 12
-
-    // Tooltip background
-    ctx.fillStyle = isDark ? 'rgba(31,30,29,0.95)' : 'rgba(255,255,255,0.96)'
-    ctx.strokeStyle = borderColor
-    ctx.lineWidth = 1
-    ctx.shadowColor = 'rgba(0,0,0,0.15)'
-    ctx.shadowBlur = 12
-    graphRoundRect(ctx, tx, ty, tw, th, 8)
-    ctx.fill()
-    ctx.stroke()
-    ctx.shadowBlur = 0
-    ctx.shadowColor = 'transparent'
-
-    ctx.fillStyle = textColor
-    ctx.font = 'bold 11px -apple-system, sans-serif'
-    ctx.textAlign = 'left'
-    ctx.fillText(text, tx + 12, ty + 18)
-    ctx.font = '10px -apple-system, sans-serif'
-    ctx.fillStyle = textMuted
-    ctx.fillText(conns, tx + 12, ty + 34)
-    if (kw) {
-      ctx.fillText(kw.length > 40 ? kw.slice(0, 40) + '...' : kw, tx + 12, ty + 50)
-    }
-  }
-
-  ctx.restore()
-}
-
-function graphRoundRect(ctx, x, y, w, h, r) {
-  ctx.beginPath()
-  ctx.moveTo(x + r, y)
-  ctx.lineTo(x + w - r, y)
-  ctx.quadraticCurveTo(x + w, y, x + w, y + r)
-  ctx.lineTo(x + w, y + h - r)
-  ctx.quadraticCurveTo(x + w, y + h, x + w - r, y + h)
-  ctx.lineTo(x + r, y + h)
-  ctx.quadraticCurveTo(x, y + h, x, y + h - r)
-  ctx.lineTo(x, y + r)
-  ctx.quadraticCurveTo(x, y, x + r, y)
-  ctx.closePath()
-}
-
-// === Graph detail panel ===
-function showGraphPanel(node) {
-  let panel = document.getElementById('graphPanel')
-  if (!panel) {
-    panel = document.createElement('div')
-    panel.id = 'graphPanel'
-    panel.className = 'graph-panel'
-    document.getElementById('memGraphView').appendChild(panel)
-  }
-  const tierLabelsMap = { hot: 'Hot', warm: 'Warm', cold: 'Cold', shared: 'Shared' }
-  const created = node.mem.created_label || ''
-  panel.innerHTML = `
-    <div class="graph-panel-header">
-      <span class="badge badge-${node.tier}">${tierLabelsMap[node.tier] || node.tier}</span>
-      <span class="graph-panel-agent">${escapeHtml(node.agent)}</span>
-      <button class="graph-panel-close" id="graphPanelCloseBtn">&times;</button>
-    </div>
-    ${created ? `<div class="graph-panel-date">${escapeHtml(created)}</div>` : ''}
-    <div class="graph-panel-content">${escapeHtml(node.mem.content)}</div>
-    <div class="graph-panel-meta">
-      ${node.keywords.length ? '<div class="graph-panel-keywords">' + node.keywords.map(k => '<span class="mem-keyword-tag">' + escapeHtml(k) + '</span>').join('') + '</div>' : ''}
-    </div>
-  `
-  panel.hidden = false
-  document.getElementById('graphPanelCloseBtn').addEventListener('click', () => {
-    graphSelectedNode = null
-    panel.hidden = true
-    renderGraph()
-  })
-}
-
-function hideGraphPanel() {
-  const panel = document.getElementById('graphPanel')
-  if (panel) panel.hidden = true
-}
-
-function openEditMemory(mem) {
-  document.getElementById('memModalTitle').textContent = t('memories.modal.title_edit')
-  document.getElementById('memAgent').value = mem.agent_id || mainAgentId()
-  document.getElementById('memTier').value = mem.tier || mem.category || 'warm'
-  document.getElementById('memContent').value = mem.content || ''
-  document.getElementById('memKeywords').value = mem.keywords || ''
-  document.getElementById('memEditId').value = mem.id
-  openModal(memModalOverlay)
-}
-
-// === Graph search integration ===
-function updateGraphSearch() {
-  const q = memSearchInput.value.trim().toLowerCase()
-  graphSearchQuery = q
-  for (const node of graphNodes) {
-    if (!q) {
-      node.searchMatch = true
-    } else {
-      const content = (node.mem.content || '').toLowerCase()
-      const kws = node.keywords.join(' ').toLowerCase()
-      const agent = (node.agent || '').toLowerCase()
-      node.searchMatch = content.includes(q) || kws.includes(q) || agent.includes(q)
-    }
-  }
-  if (graphNodes.length > 0) renderGraph()
-}
-
-// === Zoom indicator ===
-function showZoomIndicator() {
-  const el = document.getElementById('graphZoomIndicator')
-  if (!el) return
-  el.textContent = `${Math.round(graphZoom * 100)}%`
-  el.classList.add('visible')
-  clearTimeout(graphZoomIndicatorTimer)
-  graphZoomIndicatorTimer = setTimeout(() => el.classList.remove('visible'), 1200)
-}
-
-// === Graph mouse interaction (with zoom/pan) ===
-;(function initGraphInteraction() {
-  const canvas = document.getElementById('memGraphCanvas')
-  let wasDragging = false
-  let wasPanning = false
-  let mouseDownPos = { x: 0, y: 0 }
-
-  // Mouse wheel zoom
-  canvas.addEventListener('wheel', (e) => {
-    e.preventDefault()
-    const zoomFactor = e.deltaY > 0 ? 0.92 : 1.08
-    const rect = canvas.getBoundingClientRect()
-    const mx = e.clientX - rect.left
-    const my = e.clientY - rect.top
-
-    // Zoom toward cursor
-    const worldX = (mx - graphPanX) / graphZoom
-    const worldY = (my - graphPanY) / graphZoom
-
-    graphZoom = Math.max(0.3, Math.min(3.0, graphZoom * zoomFactor))
-
-    graphPanX = mx - worldX * graphZoom
-    graphPanY = my - worldY * graphZoom
-
-    showZoomIndicator()
-    if (graphNodes.length > 0) renderGraph()
-  }, { passive: false })
-
-  // Mouse move: hover detection + panning + dragging
-  canvas.addEventListener('mousemove', (e) => {
-    const rect = e.target.getBoundingClientRect()
-    const sx = e.clientX - rect.left
-    const sy = e.clientY - rect.top
-
-    // Panning
-    if (graphPanning) {
-      const dx = sx - graphPanStartX
-      const dy = sy - graphPanStartY
-      if (Math.abs(dx) > 3 || Math.abs(dy) > 3) wasPanning = true
-      graphPanX += dx
-      graphPanY += dy
-      graphPanStartX = sx
-      graphPanStartY = sy
-      if (graphNodes.length > 0) renderGraph()
-      return
-    }
-
-    // Dragging a node
-    const world = screenToWorld(sx, sy)
-    if (graphDragging) {
-      const dx = sx - mouseDownPos.x
-      const dy = sy - mouseDownPos.y
-      if (Math.abs(dx) > 3 || Math.abs(dy) > 3) wasDragging = true
-      graphDragging.x = world.x
-      graphDragging.y = world.y
-      graphDragging.vx = 0
-      graphDragging.vy = 0
-      if (graphNodes.length > 0) renderGraph()
-      return
-    }
-
-    // Hover detection in world space
-    graphHover = null
-    for (const node of graphNodes) {
-      const ndx = world.x - node.x
-      const ndy = world.y - node.y
-      const hitRadius = (node.radius + 6) / Math.max(graphZoom, 0.5)
-      if (ndx * ndx + ndy * ndy < hitRadius * hitRadius) {
-        graphHover = node
-        break
-      }
-    }
-    canvas.style.cursor = graphHover ? 'pointer' : 'grab'
-    if (graphNodes.length > 0) renderGraph()
-  })
-
-  // Mouse down: start drag on node, or start pan on empty space
-  canvas.addEventListener('mousedown', (e) => {
-    const rect = e.target.getBoundingClientRect()
-    const sx = e.clientX - rect.left
-    const sy = e.clientY - rect.top
-    mouseDownPos = { x: sx, y: sy }
-    wasDragging = false
-    wasPanning = false
-
-    if (graphHover) {
-      // Drag node
-      graphDragging = graphHover
-      canvas.style.cursor = 'grabbing'
-    } else {
-      // Pan
-      graphPanning = true
-      graphPanStartX = sx
-      graphPanStartY = sy
-      canvas.style.cursor = 'grabbing'
-    }
-  })
-
-  // Click: select node and show panel (only if not dragged/panned)
-  canvas.addEventListener('click', (e) => {
-    if (wasDragging || wasPanning) return
-
-    const rect = e.target.getBoundingClientRect()
-    const sx = e.clientX - rect.left
-    const sy = e.clientY - rect.top
-    const world = screenToWorld(sx, sy)
-
-    let clicked = null
-    for (const node of graphNodes) {
-      const dx = world.x - node.x
-      const dy = world.y - node.y
-      const hitRadius = (node.radius + 6) / Math.max(graphZoom, 0.5)
-      if (dx * dx + dy * dy < hitRadius * hitRadius) {
-        clicked = node
-        break
-      }
-    }
-
-    if (clicked) {
-      graphSelectedNode = clicked
-      showGraphPanel(clicked)
-    } else {
-      graphSelectedNode = null
-      hideGraphPanel()
-    }
-    if (graphNodes.length > 0) renderGraph()
-  })
-
-  // Double click: open edit modal
-  canvas.addEventListener('dblclick', (e) => {
-    if (graphHover && graphHover.mem) {
-      openEditMemory(graphHover.mem)
-    }
-  })
-
-  // Mouse up: stop drag/pan
-  document.addEventListener('mouseup', () => {
-    if (graphDragging) {
-      graphDragging = null
-      const c = document.getElementById('memGraphCanvas')
-      if (c) c.style.cursor = graphHover ? 'pointer' : 'grab'
-    }
-    if (graphPanning) {
-      graphPanning = false
-      const c = document.getElementById('memGraphCanvas')
-      if (c) c.style.cursor = 'grab'
-    }
-  })
-
-  // Search integration: listen to existing search input
-  memSearchInput.addEventListener('input', () => {
-    if (currentMemTier === 'graph') {
-      updateGraphSearch()
-    }
-  })
-  memSearchInput.addEventListener('keydown', (e) => {
-    if (e.key === 'Enter' && currentMemTier === 'graph') {
-      updateGraphSearch()
-    }
-  })
-})()
-
-// === Daily Log ===
-
-async function loadDailyLog() {
-  // "Minden ügynök" (empty value) falls back to the first agent in the
-  // filter dropdown, which is the main agent on any BOT_NAME -- avoids a
-  // hardcoded "marveen" slug that would 404 on zino/haver/etc installs.
-  const sel = document.getElementById('memAgentFilter')
-  const agent = sel.value || (sel.options[1] ? sel.options[1].value : '')
-  if (!agent) {
-    renderLogEntries([])
-    return
-  }
-
-  try {
-    const datesRes = await fetch(`/api/daily-log/dates?agent=${agent}`)
-    logDates = await datesRes.json()
-  } catch {
-    logDates = []
-  }
-
-  document.getElementById('logCurrentDate').textContent = formatLogDate(currentLogDate)
-
-  try {
-    const res = await fetch(`/api/daily-log?agent=${agent}&date=${currentLogDate}`)
-    const entries = await res.json()
-    renderLogEntries(entries)
-  } catch {
-    renderLogEntries([])
-  }
-}
-
-function renderLogEntries(entries) {
-  const el = document.getElementById('logEntries')
-  const empty = document.getElementById('logEmpty')
-  el.innerHTML = ''
-  empty.hidden = entries.length > 0
-
-  for (const entry of entries) {
-    const time = new Date(entry.created_at * 1000).toLocaleTimeString('hu-HU', { hour: '2-digit', minute: '2-digit' })
-    const div = document.createElement('div')
-    div.className = 'log-entry'
-    div.innerHTML = `
-      <div class="log-entry-time">${time}</div>
-      <div class="log-entry-content">${escapeHtml(entry.content)}</div>
-    `
-    el.appendChild(div)
-  }
-}
-
-function formatLogDate(dateStr) {
-  const d = new Date(dateStr + 'T12:00:00')
-  return d.toLocaleDateString('hu-HU', { year: 'numeric', month: 'long', day: 'numeric', weekday: 'long' })
-}
-
-// Date navigation
-document.getElementById('logPrevDate').addEventListener('click', () => {
-  const d = new Date(currentLogDate)
-  d.setDate(d.getDate() - 1)
-  currentLogDate = d.toISOString().split('T')[0]
-  loadDailyLog()
-})
-document.getElementById('logNextDate').addEventListener('click', () => {
-  const d = new Date(currentLogDate)
-  d.setDate(d.getDate() + 1)
-  currentLogDate = d.toISOString().split('T')[0]
-  loadDailyLog()
-})
-
-// === SVG icons ===
-function pauseIcon() {
-  return '<svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor"><rect x="6" y="4" width="4" height="16" rx="1"/><rect x="14" y="4" width="4" height="16" rx="1"/></svg>'
-}
-function playIcon() {
-  return '<svg width="14" height="14" viewBox="0 0 24 24" fill="currentColor"><polygon points="5 3 19 12 5 21 5 3"/></svg>'
-}
-function trashIcon() {
-  return '<svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><polyline points="3 6 5 6 21 6"/><path d="M19 6v14a2 2 0 01-2 2H7a2 2 0 01-2-2V6m3 0V4a2 2 0 012-2h4a2 2 0 012 2v2"/></svg>'
-}
 
 // ============================================================
 // === Connectors ===
@@ -3857,166 +2765,6 @@ async function loadCosts() {
     el.innerHTML = `<div style="${mutedStyle}">${t('costs.load_failed')}</div>`
   }
 }
-
-// ============================================================
-// === Memory Import ===
-// ============================================================
-
-const memImportOverlay = document.getElementById('memImportOverlay')
-const memImportFileInput = document.getElementById('memImportFile')
-const memImportFileArea = document.getElementById('memImportFileArea')
-const memImportFileNames = document.getElementById('memImportFileNames')
-const memImportSaveBtn = document.getElementById('memImportSaveBtn')
-const memImportProgress = document.getElementById('memImportProgress')
-const memImportStatus = document.getElementById('memImportStatus')
-const memImportResult = document.getElementById('memImportResult')
-let memImportFiles = []
-
-// Open import modal
-document.getElementById('memImportOpenBtn').addEventListener('click', () => {
-  memImportFiles = []
-  memImportFileInput.value = ''
-  memImportFileNames.textContent = ''
-  memImportProgress.hidden = true
-  memImportResult.hidden = true
-  memImportSaveBtn.querySelector('.btn-text').hidden = false
-  memImportSaveBtn.querySelector('.btn-loading').hidden = true
-  memImportSaveBtn.disabled = false
-
-  // Populate agent dropdown from existing agents
-  const importAgentSel = document.getElementById('memImportAgent')
-  const memAgentSel = document.getElementById('memAgent')
-  importAgentSel.innerHTML = memAgentSel.innerHTML
-  openModal(memImportOverlay)
-})
-
-// Close import modal
-document.getElementById('memImportClose').addEventListener('click', () => closeModal(memImportOverlay))
-memImportOverlay.addEventListener('click', (e) => { if (e.target === memImportOverlay) closeModal(memImportOverlay) })
-
-// File area click -> trigger file input
-memImportFileArea.addEventListener('click', () => memImportFileInput.click())
-
-// Drag and drop
-memImportFileArea.addEventListener('dragover', (e) => {
-  e.preventDefault()
-  memImportFileArea.style.borderColor = 'var(--accent)'
-})
-memImportFileArea.addEventListener('dragleave', () => {
-  memImportFileArea.style.borderColor = ''
-})
-memImportFileArea.addEventListener('drop', (e) => {
-  e.preventDefault()
-  memImportFileArea.style.borderColor = ''
-  const files = Array.from(e.dataTransfer.files).filter(f =>
-    f.name.endsWith('.md') || f.name.endsWith('.txt') || f.name.endsWith('.json')
-  )
-  if (files.length) {
-    memImportFiles = files
-    memImportFileNames.textContent = files.map(f => f.name).join(', ')
-  }
-})
-
-// File input change
-memImportFileInput.addEventListener('change', () => {
-  memImportFiles = Array.from(memImportFileInput.files)
-  memImportFileNames.textContent = memImportFiles.map(f => f.name).join(', ')
-})
-
-// Parse file into chunks (client-side)
-async function parseFileToChunks(file) {
-  const text = await file.text()
-  const ext = file.name.split('.').pop().toLowerCase()
-
-  if (ext === 'json') {
-    try {
-      const data = JSON.parse(text)
-      if (Array.isArray(data)) {
-        return data.map(item => {
-          if (typeof item === 'object' && item !== null) return item.content || item.text || item.value || JSON.stringify(item)
-          return String(item)
-        }).filter(s => s.length > 20).map(s => s.slice(0, 2000))
-      }
-      return Object.entries(data).map(([k, v]) => `${k}: ${v}`).filter(s => s.length > 20).map(s => s.slice(0, 2000))
-    } catch { return [text.slice(0, 2000)] }
-  }
-
-  if (ext === 'md') {
-    return text.split(/\n(?=##?\s)/).map(s => s.trim()).filter(s => s.length > 20).map(s => s.slice(0, 2000))
-  }
-
-  // txt: split by paragraphs
-  return text.split(/\n\n+/).map(s => s.trim()).filter(s => s.length > 20).map(s => s.slice(0, 2000))
-}
-
-// Import button click
-memImportSaveBtn.addEventListener('click', async () => {
-  if (!memImportFiles.length) {
-    showToast(t('memories.toast.select_files'))
-    return
-  }
-
-  memImportSaveBtn.querySelector('.btn-text').hidden = true
-  memImportSaveBtn.querySelector('.btn-loading').hidden = false
-  memImportSaveBtn.disabled = true
-  memImportProgress.hidden = false
-  memImportResult.hidden = true
-  memImportStatus.textContent = t('memories.import.processing')
-
-  try {
-    // Parse all files into chunks
-    let allChunks = []
-    for (const file of memImportFiles) {
-      const chunks = await parseFileToChunks(file)
-      allChunks = allChunks.concat(chunks)
-    }
-
-    if (allChunks.length === 0) {
-      memImportProgress.hidden = true
-      memImportSaveBtn.querySelector('.btn-text').hidden = false
-      memImportSaveBtn.querySelector('.btn-loading').hidden = true
-      memImportSaveBtn.disabled = false
-      showToast(t('memories.toast.no_content'))
-      return
-    }
-
-    memImportStatus.textContent = t('memories.import.importing', { n: allChunks.length })
-
-    const agentId = document.getElementById('memImportAgent').value || mainAgentId()
-    const resp = await fetch('/api/memories/import', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ agent_id: agentId, chunks: allChunks }),
-    })
-    const data = await resp.json()
-
-    memImportProgress.hidden = true
-
-    if (data.ok) {
-      const s = data.stats || {}
-      memImportResult.hidden = false
-      memImportResult.innerHTML = `
-        <div style="color:var(--text-primary);font-weight:600;margin-bottom:8px">${t('memories.import.done_title')}</div>
-        <div style="font-size:13px;color:var(--text-secondary)">
-          ${t('memories.import.done_sub', { n: `<strong>${data.imported}</strong>` })}<br>
-          Hot: ${s.hot || 0} | Warm: ${s.warm || 0} | Cold: ${s.cold || 0} | Shared: ${s.shared || 0}
-        </div>
-      `
-      showToast(t('memories.toast.imported', { n: data.imported }))
-      loadMemories()
-      loadMemStats()
-    } else {
-      showToast('Hiba: ' + (data.error || 'Ismeretlen'))
-    }
-  } catch (err) {
-    memImportProgress.hidden = true
-    showToast(t('memories.toast.import_error'))
-  }
-
-  memImportSaveBtn.querySelector('.btn-text').hidden = false
-  memImportSaveBtn.querySelector('.btn-loading').hidden = true
-  memImportSaveBtn.disabled = false
-})
 
 // ============================================================
 // === Költöztetés (Migration) ===
